dev: merge `TaskWhen` and move `PathPattern`
